ci_scripts/requirements.txt
gitpython
towncrier
pyautoversion
boto3
requests >= 2.20
python-dotenv
jinja2
mbed-os-tools
toml
twine<1.12
